# packages/contracts/openapi/admin-api.yaml
openapi: 3.1.0
info:
  title: Admin API
  version: 0.4.4
  description: |
    API behind the admin application (Store view + HQ view). Every request carries a Keycloak staff-realm JWT.
    Store-scoped resources live under `/admin/stores/{storeId}/...`; the server checks the OpenFGA relation named
    in each operation's `x-permission` against `store:{storeId}` (or `organization:hq`) before doing anything.
    UI gating is a convenience; these checks are the security boundary. Every mutation writes `audit_log`.
    Money is `{ amount_minor, currency }`. Ids are uuids. Frozen at tag contracts-v0.1; 0.2.0 adds optional sort/order to list operations (additive);
    0.3.0 adds the marketing surface under `/admin/stores/{storeId}/marketing/…` and `/admin/marketing/…` (docs/marketing-scope.md, additive).
    0.4.0 adds search merchandising rules under `/admin/stores/{storeId}/merchandising/…` (CONTRACT CHANGE #162, additive) and
    documents `401` and `403` on every operation (#180): every `x-permission` can fail, and Prism can only mock what is written down.
    0.4.1 adds per-item product media + signed Cloudinary upload params (#168), `getPromotion`/`updatePromotion` with the
    `buy_x_get_y` type and `stackable`/`exclusive` (#189), and spells `ProductFeed` out so `status: error` validates (#194).
servers:
  - url: http://localhost:4011
    description: Prism mock (pnpm mock)
  - url: https://admin-api.example.com/v1
security:
  - staffToken: []
...
paths:
...
  /admin/stores/{storeId}/marketing/campaigns:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
    get:
      tags: [marketing]
      operationId: listCampaigns
      x-permission: { relation: store_staff, object: 'store:{storeId}' }
      parameters:
        - {
            name: status,
            in: query,
            schema: { type: string, enum: [draft, scheduled, active, paused, ended] },
          }
        - {
            name: type,
            in: query,
            schema:
              {
                type: string,
                enum: [email, sms, paid_social, paid_search, affiliate, referral, landing],
              },
          }
        - {
            name: sort,
            in: query,
            schema:
              { type: string, enum: [name, status, starts_at, created_at], default: created_at },
          }
        - { $ref: '#/components/parameters/Order' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/Campaign' } }
              example:
                page: 1
                limit: 20
                total: 1
                items:
                  - id: 70000000-0000-4000-8000-000000000701
                    store_id: 00000000-0000-4000-8000-000000000031
                    name: Autumn launch
                    type: paid_social
                    status: draft
                    starts_at: '2026-09-15T00:00:00Z'
                    ends_at: '2026-10-15T00:00:00Z'
                    budget: { amount_minor: 250000, currency: EUR }
                    utm_source: meta
                    utm_medium: paid_social
                    utm_campaign: autumn-2026
                    promotion_id: 50000000-0000-4000-8000-000000000101
                    segment_id: null
                    landing_path: /collections/autumn
                    external_ref: '120210000000000001'
                    launched_at: null
                    ended_at: null
                    metadata: {}
                    created_at: '2026-09-08T00:00:00Z'
                    updated_at: '2026-09-08T00:00:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [marketing]
      operationId: createCampaign
      summary: Create a campaign in status draft
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CampaignInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
              examples:
                autumn: { $ref: '#/components/examples/CampaignAutumn' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
  /admin/stores/{storeId}/marketing/campaigns/{campaignId}:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: campaignId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [marketing]
      operationId: getCampaign
      x-permission: { relation: store_staff, object: 'store:{storeId}' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
              examples:
                autumn: { $ref: '#/components/examples/CampaignAutumn' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [marketing]
      operationId: updateCampaign
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CampaignInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
              examples:
                autumn: { $ref: '#/components/examples/CampaignAutumn' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [marketing]
      operationId: deleteCampaign
      summary: Delete a draft campaign (409 once launched; end it instead)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/stores/{storeId}/marketing/campaigns/{campaignId}/launch:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: campaignId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [marketing]
      operationId: launchCampaign
      summary: draft/scheduled/paused → active (emits campaign.launched)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
              examples:
                autumn: { $ref: '#/components/examples/CampaignAutumnActive' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/stores/{storeId}/marketing/campaigns/{campaignId}/end:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: campaignId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [marketing]
      operationId: endCampaign
      summary: active/paused → ended (emits campaign.ended)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Campaign' }
              example:
                id: 70000000-0000-4000-8000-000000000701
                store_id: 00000000-0000-4000-8000-000000000031
                name: Autumn launch
                type: paid_social
                status: ended
                starts_at: '2026-09-15T00:00:00Z'
                ends_at: '2026-10-15T00:00:00Z'
                budget: { amount_minor: 250000, currency: EUR }
                utm_source: meta
                utm_medium: paid_social
                utm_campaign: autumn-2026
                promotion_id: 50000000-0000-4000-8000-000000000101
                segment_id: null
                landing_path: /collections/autumn
                external_ref: '120210000000000001'
                launched_at: '2026-09-15T00:00:00Z'
                ended_at: '2026-10-15T00:00:00Z'
                metadata: {}
                created_at: '2026-09-08T00:00:00Z'
                updated_at: '2026-10-15T00:00:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/stores/{storeId}/marketing/segments:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
    get:
      tags: [marketing]
      operationId: listSegments
      x-permission: { relation: store_staff, object: 'store:{storeId}' }
      parameters:
        - {
            name: sort,
            in: query,
            schema:
              {
                type: string,
                enum: [name, materialised_count, last_materialised_at, created_at],
                default: created_at,
              },
          }
        - { $ref: '#/components/parameters/Order' }
        - { $ref: '#/components/parameters/Page' }
        - { $ref: '#/components/parameters/Limit' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Page'
                  - type: object
                    required: [items]
                    properties:
                      items: { type: array, items: { $ref: '#/components/schemas/Segment' } }
              example:
                page: 1
                limit: 20
                total: 1
                items:
                  - id: 70000000-0000-4000-8000-000000000721
                    store_id: 00000000-0000-4000-8000-000000000031
                    name: VIP
                    description: Spent 500+ in the last year and opted in to email
                    rules:
                      v: 1
                      all:
                        - any: [{ field: total_spent_minor, op: gte, value: 50000 }]
                        - any: [{ field: last_order_at, op: after, value: '2025-09-01T00:00:00Z' }]
                        - any: [{ field: consent, op: granted, value: email }]
                    template_id: 70000000-0000-4000-8000-000000000711
                    materialised_count: 128
                    last_materialised_at: '2026-09-08T03:00:00Z'
                    created_at: '2026-09-08T00:00:00Z'
                    updated_at: '2026-09-08T03:00:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    post:
      tags: [marketing]
      operationId: createSegment
      summary: Create a segment (optionally from an organization template via template_id)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SegmentInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Segment' }
              examples:
                vip: { $ref: '#/components/examples/SegmentVip' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/stores/{storeId}/marketing/segments/{segmentId}:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: segmentId, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [marketing]
      operationId: getSegment
      x-permission: { relation: store_staff, object: 'store:{storeId}' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Segment' }
              examples:
                vip: { $ref: '#/components/examples/SegmentVip' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [marketing]
      operationId: updateSegment
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SegmentInput' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Segment' }
              examples:
                vip: { $ref: '#/components/examples/SegmentVip' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
    delete:
      tags: [marketing]
      operationId: deleteSegment
      summary: Delete a segment and its members (409 while a campaign references it)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/stores/{storeId}/marketing/segments/{segmentId}/preview:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: segmentId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [marketing]
      operationId: previewSegment
      summary: Count the customers the saved rules (or the rules in the body) match right now; writes nothing
      x-permission: { relation: store_staff, object: 'store:{storeId}' }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                rules: { $ref: '#/components/schemas/SegmentRules' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [count]
                properties:
                  count: { type: integer, minimum: 0 }
              example: { count: 128 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/stores/{storeId}/marketing/segments/{segmentId}/materialize:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: segmentId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [marketing]
      operationId: materializeSegment
      summary: Refresh segment_member from the rules (async job; materialised_count updates when it finishes)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Segment' }
              examples:
                vip: { $ref: '#/components/examples/SegmentVip' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
  /admin/stores/{storeId}/marketing/feeds/{feedId}/publish:
    parameters:
      - { $ref: '#/components/parameters/StoreId' }
      - { name: feedId, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [marketing]
      operationId: publishFeed
      summary: Generate the feed file and serve it at `url` (emits feed.published; status → active or error)
      x-permission: { relation: store_admin, object: 'store:{storeId}' }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductFeed' }
              examples:
                google: { $ref: '#/components/examples/FeedGoogle' }
                googleError: { $ref: '#/components/examples/FeedGoogleError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
...
components:
  securitySchemes:
    staffToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Keycloak staff realm token
  parameters:
    StoreId:
      name: storeId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Order:
      name: order
      in: query
      description: Sort direction. Ignored unless `sort` is present.
      schema: { type: string, enum: [asc, desc], default: desc }
  responses:
...
    Unauthorized:
      description: Missing or invalid token
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: unauthorized, message: Missing bearer token, details: {} }
    Forbidden:
      description: Permission check failed (OpenFGA)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example:
            {
              code: forbidden,
              message: 'requires finance on organization:hq',
              details: { relation: finance, object: 'organization:hq' },
            }
    NotFound:
      description: Not found in the caller's scope
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: not_found, message: Not found, details: {} }
...
  schemas:
...
    Page:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
...